+++ localalias/app.py
@@ -82,7 +82,7 @@ def _get_argparser():
     )
     command_group.add_argument(
         '-R', nargs=2, dest='command_args', action=_CmdAction, metavar=('OLD', 'NEW'),
-        help='Rename an existing local alias. OLD is renamed to NEW.'
+        help='Rename an existing local alias. OLD alias is renamed to NEW.'
     )
     command_group.add_argument(
         'command_args', nargs='?', action=_CmdAction, metavar='PREFIX',
@@ -101,7 +101,7 @@ class _CmdAction(argparse.Action):
 
     def __call__(self, parser, namespace, values, option_string):
         if self.__class__.flag is None:
-            self.__class__.flag = self._flag_map(option_string)
+            self.__class__.flag = _CmdFlag(option_string)
             self.__class__.option_string = option_string
         elif option_string is not None:
             raise errors.ArgumentError(
@@ -133,21 +133,12 @@ class _CmdAction(argparse.Action):
         cmd = cmd_builder(args.command_args, color=args.color)
         return cmd()
 
-    def _flag_map(self, option_string):
-        """Maps option strings to Command Flags"""
-        return {'-a': _CmdFlag.ADD,
-                '-r': _CmdFlag.REMOVE,
-                '-e': _CmdFlag.EDIT,
-                '-x': _CmdFlag.EXECUTE,
-                '-R': _CmdFlag.RENAME,
-                None: _CmdFlag.SHOW}[option_string]
-
 
 class _CmdFlag(enum.Enum):
     """Command Flags"""
-    ADD = 1
-    REMOVE = 2
-    EDIT = 3
-    EXECUTE = 4
-    SHOW = 5
-    RENAME = 6
+    ADD = '-a'
+    REMOVE = '-r'
+    EDIT = '-e'
+    EXECUTE = '-x'
+    SHOW = None
+    RENAME = '-R'